chore: Move Blog collections to dedicated directory
