+++ apps/civicsignalblog/src/payload/globals/Site/explorer.js
@@ -4,11 +4,11 @@ import EngagementTab from "./Components/EngagementTab";
 import GeneralTab from "./Components/GeneralTab";
 import NavigationTab from "./Components/NavigationTab";
 
-const Explorer = createSettings(
-  "settings-explorer-site",
-  "Explorer Site",
-  "Settings",
-  [GeneralTab, NavigationTab, EngagementTab],
-);
+const Explorer = createSettings({
+  slug: "settings-explorer-site",
+  label: "Explorer Site",
+  group: "Settings",
+  tabs: [GeneralTab, NavigationTab, EngagementTab],
+});
 
 export default Explorer;
