+++ apps/civicsignalblog/src/payload/globals/Site/research.js
@@ -4,10 +4,11 @@ import EngagementTab from "./Components/EngagementTab";
 import GeneralTab from "./Components/GeneralTab";
 import NavigationTab from "./Components/NavigationTab";
 
-const Site = createSettings("settings-site", "Research Site", "Settings", [
-  GeneralTab,
-  NavigationTab,
-  EngagementTab,
-]);
+const Site = createSettings({
+  slug: "settings-site",
+  label: "Research Site",
+  group: "Settings",
+  tabs: [GeneralTab, NavigationTab, EngagementTab],
+});
 
 export default Site;
