+++ apps/civicsignalblog/src/payload/globals/Site/sources.js
@@ -4,11 +4,11 @@ import EngagementTab from "./Components/EngagementTab";
 import GeneralTab from "./Components/GeneralTab";
 import NavigationTab from "./Components/NavigationTab";
 
-const Sources = createSettings(
-  "settings-sources-site",
-  "Source Manager Site",
-  "Settings",
-  [GeneralTab, NavigationTab, EngagementTab],
-);
+const Sources = createSettings({
+  slug: "settings-sources-site",
+  label: "Source Manager Site",
+  group: "Settings",
+  tabs: [GeneralTab, NavigationTab, EngagementTab],
+});
 
 export default Sources;
