+++ apps/civicsignalblog/src/payload/globals/Site/topicMapper.js
@@ -4,11 +4,11 @@ import EngagementTab from "./Components/EngagementTab";
 import GeneralTab from "./Components/GeneralTab";
 import NavigationTab from "./Components/NavigationTab";
 
-const TopicMapper = createSettings(
-  "settings-topic-mapper-site",
-  "Topic Mapper Site",
-  "Settings",
-  [GeneralTab, NavigationTab, EngagementTab],
-);
+const TopicMapper = createSettings({
+  slug: "settings-topic-mapper-site",
+  label: "Topic Mapper Site",
+  group: "Settings",
+  tabs: [GeneralTab, NavigationTab, EngagementTab],
+});
 
 export default TopicMapper;
